refactor: mofify helper function
