+++ src/helper/main/document/removeTag.js
