+++ src/helper/main/setViewRender.js
