+++ src/pages/Main.js
@@ -8,8 +8,8 @@ import useDependency from "../hooks/useDependency";
 import javascriptSvg from "../assets/images/javascript.svg";
 import reactSvg from "../assets/images/react.svg";
 
-import { updateFileContent } from "../helper/searchDfs";
-import { setViewRender } from "../helper/setViewRender";
+import { updateFileContent } from "../helper/main/fileTreeHelper";
+import { setViewRender } from "../helper/main/setViewRender";
 
 import Button from "../components/common/Button";
